SIFRPC: Move code to SIF file
